Modified host to be accessible outside of the local machine
Modified itemsApi, luggagesApi, tripsApi to use the local machines IP. This is an example and will no longer be valid if the local machine's IP changes.

Modified package.json to host the backend server on the same IP address. Again, this must be changed if the machine's IP changes or if it is hosted on another device (with a different IP)